+++ src/file_span.rs
@@ -1,0 +1,68 @@
+use crate::src_file::SrcID;
+
+#[derive(Clone, Copy, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub struct FileSpan {
+  pub id: SrcID,
+  pub start: usize,
+  pub end: usize,
+}
+impl chumsky::span::Span for FileSpan {
+  type Offset = usize;
+  type Context = SrcID;
+  #[inline]
+  #[must_use]
+  fn new(context: Self::Context, range: std::ops::Range<Self::Offset>) -> Self {
+    Self { id: context, start: range.start, end: range.end }
+  }
+  #[inline]
+  #[must_use]
+  fn start(&self) -> Self::Offset {
+    self.start
+  }
+  #[inline]
+  #[must_use]
+  fn end(&self) -> Self::Offset {
+    self.end
+  }
+  #[inline]
+  #[must_use]
+  fn context(&self) -> Self::Context {
+    self.id
+  }
+}
+impl ariadne::Span for FileSpan {
+  type SourceId = SrcID;
+
+  #[inline]
+  #[must_use]
+  fn source(&self) -> &Self::SourceId {
+    &self.id
+  }
+
+  #[inline]
+  #[must_use]
+  fn start(&self) -> usize {
+    self.start
+  }
+
+  #[inline]
+  #[must_use]
+  fn end(&self) -> usize {
+    self.end
+  }
+}
+impl core::fmt::Debug for FileSpan {
+  #[inline]
+  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    core::fmt::Display::fmt(self, f)
+  }
+}
+impl core::fmt::Display for FileSpan {
+  #[inline]
+  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    let info = self.id.get_src_file();
+    let path = info.path();
+    let (line, col) = info.line_col(self.start);
+    write!(f, "{path}:{line}:{col}", path = path.display())
+  }
+}
