+++ src/lib.rs
@@ -14,8 +14,9 @@
 //!   reading right now) is **NOT** covered by the semantic version of the
 //!   package.
 
-pub mod asm;
 pub mod ast;
 pub mod errors;
-pub mod src_files;
+pub mod file_span;
+pub mod file_spanned;
+pub mod src_file;
 pub mod str_id;
